inclusão do nome do livro no alerta de sucesso.

+++ controllers/bookControllers.js
@@ -5,8 +5,8 @@ import { ObjectId } from "mongodb";
 const createNewBook = async (req, res) => {
     try {
         const { title, author, yearPublication, description } = req.body;
-        await bookService.create(title, author, yearPublication, description);
-        res.status(201).json({ Success: 'Livro cadastrado com sucesso' }); //Cód. Status 201: Create
+        const newBook = await bookService.create(title, author, yearPublication, description);
+        res.status(201).json({ Success: `Livro '${newBook.title}' cadastrado com sucesso` }); //Cód. Status 201: Create
     } catch (error) {
         console.log(error);
         res.status(500).json({ error: 'Erro interno do servidor' }); //Cód. Status 500: Internal Server Error
@@ -45,8 +45,10 @@ const updateBook = async (req, res) => {
         if(ObjectId.isValid(req.params.id)) {
             const id = req.params.id;
             const { title, author, yearPublication, description } = req.body;
+            const existingBook = await bookService.getOne(id);
+            if (!existingBook) { return res.status(404).json({ error: 'Livro não encontrado.' }); }
             bookService.update(id, title, author, yearPublication, description);
-            res.status(200).json({ Success: 'Alteração feita com sucesso.' }); //Cód. Status 200: OK
+            res.status(200).json({ Success: `Livro '${existingBook.title}' atualizado com sucesso.` }); //Cód. Status 200: OK
         } else { res.sendStatus(400); } //Cód. Status 400: Bad Request
     } catch (error) {
         console.log(error);
@@ -59,8 +61,10 @@ const deleteBook = async (req, res) => {
     try {
         if(ObjectId.isValid(req.params.id)) {
             const id = req.params.id;
+            const existingBook = await bookService.getOne(id);
+            if (!existingBook) { return res.status(404).json({ error: 'Livro não encontrado.' }); }
             bookService.delete(id);
-            res.status(200).json({ Success: 'Livro deletado com sucesso.' }); //Cód. Status 200: OK
+            res.status(200).json({ Success: `Livro '${existingBook.title}' deletado com sucesso.` }); //Cód. Status 200: OK
         } else {res.sendStatus(400); } //Cód. Status 400: Bad Request
     } catch (error) {
         console.log(error);
